move the handle_choice and run methods in main.rb
