+++ app.rb
@@ -4,8 +4,7 @@ require_relative 'teacher'
 require_relative 'book'
 require_relative 'rentals'
 require_relative 'modules'
-require_relative 'homepage'
-require_relative 'create_teacher'
+require_relative 'create_teacher_student'
 
 class App
   include LibraryUtils
@@ -13,30 +12,6 @@ class App
     @books = []
     @people = []
     @rentals = []
-  end
-
-  def run
-    puts 'Welcome to the Console Library App!'
-    loop do
-      display_options
-      choice = gets.chomp.to_i
-      handle_choice(choice)
-      break if choice == 7
-    end
-    puts 'Thank you for using the Console Library App. Goodbye!'
-  end
-
-  def handle_choice(choice)
-    case choice
-    when 1 then list_all_books
-    when 2 then list_all_people
-    when 3 then create_person
-    when 4 then create_book
-    when 5 then create_rental
-    when 6 then list_rentals_for_person
-    else
-      puts 'Invalid option. Please try again.'
-    end
   end
 
   def create_person
@@ -52,20 +27,6 @@ class App
       create_teacher(name, age)
     else
       create_student(name, age)
-    end
-  end
-
-  def create_student(name, age)
-    print 'Has parent permission? [Y/N] : '
-    permission = gets.chomp
-    if permission == 'Y'
-      student = Student.new(age, 'first', name)
-      classroom = Classroom.new('first')
-      classroom.add_student(student)
-      @people << student
-      puts 'Person created successfully.'
-    else
-      puts 'No permission = No account creation'
     end
   end
